requirements.txt
# Lane MVP Dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
PyYAML>=6.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
matplotlib>=3.7.0
# bag 解码 (按实际 bag 格式选装)
# rosbag>=1.16.0        # ROS1 bag
# rosbags>=0.9.0        # ROS2 bag
